+++ server/irmaserver/sessions.go
@@ -107,7 +107,7 @@ var (
 
 	minProtocolVersion       = irma.NewVersion(2, 4)
 	minSecureProtocolVersion = irma.NewVersion(2, 8)
-	maxProtocolVersion       = irma.NewVersion(2, 8)
+	maxProtocolVersion       = irma.NewVersion(2, 9)
 
 	minFrontendProtocolVersion = irma.NewVersion(1, 0)
 	maxFrontendProtocolVersion = irma.NewVersion(1, 1)